feat: thêm chức năng thêm hàng hóa cho trang hàng hóa và đơn vị của chủ cửa hàng.

+++ backend/src/BizFlow.Domain/Entities/InventoryTransaction.cs
@@ -9,7 +9,7 @@ public class InventoryTransaction
     public Guid TenantId { get; set; }
     public Guid ProductId { get; set; }
     public InventoryTransactionType Type { get; set; }
-    public int Quantity { get; set; } // Số lượng thay đổi (tính theo base_unit)
+    public decimal Quantity { get; set; } // Số lượng thay đổi (tính theo base_unit)
     public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
     public Guid? CreatedBy { get; set; }
     public string? Note { get; set; }
